Fix potential bug in c.comp startNodes assignment
In short, the previous version had a relatively small chance of
not assigning a startNode for certain connected components.

Now, the code that assigns to startNodes iterates over every node
in the graph -- ensuring that each node's corresponding connected
component has a startNode assigned. So this should prevent the
aforementioned problem.